HAL-1119: Allow reloading a standalone server from reload dialog
