+++ gui/src/main/java/org/jboss/as/console/client/core/UIConstants.properties
@@ -182,6 +182,7 @@ failedToReadMetricDescription=Failed to retrieve metric descriptions.
 failedToRemoveServer=Failed to remove server
 failedToResolveExpression=Failed to resolve expression
 failedToRetrieveAttributeDescriptions=Failed to retrieve attribute descriptions.
+goToRuntime=Go to Runtime
 groupAlreadyExists=Group already exists
 help=Help
 help_close_help=Close
@@ -405,7 +406,7 @@ properties_global_desc=These properties are available throughout the configurati
 reason=Reason
 reloadHost=Reload Host
 reloadServerGroup=Reload Server Group
-reloadServerNow=Go to Runtime
+reloadServerNow=Reload server Now!
 remoteSocketDescription=Configuration for outbound socket bindings for remote destinations.
 removeProfile=Remove Profile
 replaceDeployment=Replace Deployment
@@ -439,7 +440,8 @@ server_config_desc=Server configuration status. In some cases the configuration 
 server_config_uptodate=The server configuration is up to date!
 server_instance_pleaseSelect=Please select a server instance:
 server_instance_reloadRequired=A server reload is required!
-server_instance_servers_needReload=The following servers need to be reloaded from the Runtime tab:
+server_instance_servers_needReload=The following servers need to be reloaded:
+server_instance_servers_needReload_from_runtime=The following servers need to be reloaded from the Runtime tab:
 server_instance_servers_needRestart=The following servers need to be restarted from the Runtime tab:
 server_reload_desc=The server configuration needs to be reloaded or the server restarted!
 server_reload_title=Reload server configuration
