+++ gui/src/main/java/org/jboss/as/console/client/core/UIConstants_de.properties
@@ -182,6 +182,7 @@ failedToReadMetricDescription=Abrufen der metrischen Beschreibungen fehlgeschlag
 failedToRemoveServer=Entfernen von Server fehlgeschlagen
 failedToResolveExpression=Aufl\u00F6sen des Ausdrucks fehlgeschlagen
 failedToRetrieveAttributeDescriptions=Abrufen der Attributbeschreibungen fehlgeschlagen.
+goToRuntime=Go to Runtime
 groupAlreadyExists=Gruppe existiert bereits
 help=Hilfe
 help_close_help=Schlie\u00DFen
@@ -405,7 +406,7 @@ properties_global_desc=Diese Properties sind in der Konfiguration durchweg verf\
 reason=Grund
 reloadHost=Reload Host
 reloadServerGroup=Servergruppe neu laden
-reloadServerNow=Server jetzt neu laden\!
+reloadServerNow=Reload server Now!
 remoteSocketDescription=Konfigurationsinformationen f\u00FCr ein ausgehendes Socket-Binding f\u00FCr ein Remote-Ziel.
 removeProfile=Profil entfernen
 replaceDeployment=Deployment ersetzen
@@ -439,7 +440,8 @@ server_config_desc=Serverkonfigurationsstatus. In einigen F\u00E4llen muss die K
 server_config_uptodate=Die Serverkonfiguration ist auf dem neuesten Stand\!
 server_instance_pleaseSelect=Bitte eine Serverinstanz w\u00E4hlen\:
 server_instance_reloadRequired=Ein erneutes Laden des Servers ist erforderlich\!
-server_instance_servers_needReload=Die folgenden Server m\u00FCssen neu geladen werden\:
+server_instance_servers_needReload=The following servers need to be reloaded:
+server_instance_servers_needReload_from_runtime=Die folgenden Server m\u00FCssen neu geladen werden\:
 server_instance_servers_needRestart=Die folgenden Server m\u00FCssen neu gestartet werden\:
 server_reload_desc=Die Serverkonfiguration muss neu geladen oder der Server neu gestartet werden\!
 server_reload_title=Serverkonfiguration neu laden
