+++ gui/src/main/java/org/jboss/as/console/client/core/UIConstants_es.properties
@@ -182,6 +182,7 @@ failedToReadMetricDescription=No se logr\u00F3 recuperar las descripciones de la
 failedToRemoveServer=No se logr\u00F3 borrar el servidor
 failedToResolveExpression=No se logr\u00F3 resolver la expresi\u00F3n
 failedToRetrieveAttributeDescriptions=No se logr\u00F3 recuperar las descripciones de los atributos.
+goToRuntime=Go to Runtime
 groupAlreadyExists=El grupo ya existe
 help=Ayuda
 help_close_help=Cerrar
@@ -405,7 +406,7 @@ properties_global_desc=Estas propiedades est\u00E1n disponibles a trav\u00E9s de
 reason=Raz\u00F3n
 reloadHost=Reload Host
 reloadServerGroup=Re-cargar el grupo de servidores
-reloadServerNow=\u00A1Re-cargar el(los) servidor(es) ya\!
+reloadServerNow=Reload server Now!
 remoteSocketDescription=Informaci\u00F3n de configuraci\u00F3n para un destino remoto o un enlace de socket saliente.
 removeProfile=Borrar perfil
 replaceDeployment=Reemplazar implementaci\u00F3n
@@ -439,7 +440,8 @@ server_config_desc=Estatus de configuraci\u00F3n del servidor. En algunos casos 
 server_config_uptodate=\u00A1La configuraci\u00F3n del servidor parece actualizada\!
 server_instance_pleaseSelect=Seleccione una instancia del servidor\:
 server_instance_reloadRequired=\u00A1Se necesita volver a cargar el servidor\!
-server_instance_servers_needReload=Es necesario volver a cargar los siguientes servidores\:
+server_instance_servers_needReload=The following servers need to be reloaded:
+server_instance_servers_needReload_from_runtime=Es necesario volver a cargar los siguientes servidores\:
 server_instance_servers_needRestart=Es necesario reiniciar los siguientes servidores\:
 server_reload_desc=\u00A1Es necesario volver a cargar la configuraci\u00F3n del servidor o reinicar el servidor\!
 server_reload_title=Re-cargar la configuraci\u00F3n del servidor
