+++ gui/src/main/java/org/jboss/as/console/client/core/UIConstants_fr.properties
@@ -182,6 +182,7 @@ failedToReadMetricDescription=N\u0027a pas pu extraire les descriptions de m\u00
 failedToRemoveServer=N\u0027a pas pu supprimer le serveur
 failedToResolveExpression=N\u0027a pas pu r\u00E9soudre l\u0027expression
 failedToRetrieveAttributeDescriptions=N\u0027a pas pu extraire les descriptions d\u0027attribut
+goToRuntime=Go to Runtime
 groupAlreadyExists=Le groupe existe d\u00E9j\u00E0
 help=Assistance
 help_close_help=Fermer
@@ -405,7 +406,7 @@ properties_global_desc=Ces propri\u00E9t\u00E9s sont disponibles dans la configu
 reason=Raison
 reloadHost=Reload Host
 reloadServerGroup=Chargez \u00E0 nouveau de groupe de serveurs
-reloadServerNow=Chargez \u00E0 nouveau le ou les serveur(s) maintenant
+reloadServerNow=Reload server Now!
 remoteSocketDescription=Information de configuration pour une liaison de socket sortant, destination distante.
 removeProfile=Supprimer Profil
 replaceDeployment=Remplacer D\u00E9ploiement
@@ -439,7 +440,8 @@ server_config_desc=Statut de la configuration du serveur. Dans certains cas, la 
 server_config_uptodate=La configuration du serveur est mise \u00E0 jour
 server_instance_pleaseSelect=Veuillez s\u00E9lectionner une instance de serveur \:
 server_instance_reloadRequired=Un rechargement du serveur s\u0027impose.
-server_instance_servers_needReload=Les serveurs suivants ont besoin d\u0027\u00EAtre charg\u00E9s \u00E0 nouveau \:
+server_instance_servers_needReload=The following servers need to be reloaded:
+server_instance_servers_needReload_from_runtime=Les serveurs suivants ont besoin d\u0027\u00EAtre charg\u00E9s \u00E0 nouveau \:
 server_instance_servers_needRestart=Le serveurs suivants ont besoin d\u0027\u00EAtre d\u00E9marr\u00E9s \u00E0 nouveau \:
 server_reload_desc=La configuration du serveur doit \u00EAtre charg\u00E9e \u00E0 nouveau ou le serveur doit \u00EAtre d\u00E9marr\u00E9 \u00E0 nouveau.
 server_reload_title=T\u00E9l\u00E9charger \u00E0 nouveau la configuration du serveur
