+++ gui/src/main/java/org/jboss/as/console/client/core/UIConstants_ja.properties
@@ -182,6 +182,7 @@ failedToReadMetricDescription=\u6E2C\u5B9A\u57FA\u6E96\u306E\u8A73\u7D30\u3092\u
 failedToRemoveServer=\u30B5\u30FC\u30D0\u30FC\u3092\u524A\u9664\u3067\u304D\u307E\u305B\u3093\u3067\u3057\u305F
 failedToResolveExpression=\u5F0F\u3092\u89E3\u6C7A\u3067\u304D\u307E\u305B\u3093\u3067\u3057\u305F
 failedToRetrieveAttributeDescriptions=\u5C5E\u6027\u306E\u8A73\u7D30\u3092\u8AAD\u307F\u51FA\u305B\u307E\u305B\u3093\u3067\u3057\u305F\u3002
+goToRuntime=Go to Runtime
 groupAlreadyExists=\u30B0\u30EB\u30FC\u30D7\u306F\u3059\u3067\u306B\u5B58\u5728\u3057\u307E\u3059
 help=\u30D8\u30EB\u30D7
 help_close_help=\u7D42\u4E86
@@ -405,7 +406,7 @@ properties_global_desc=\u3053\u308C\u3089\u306E\u30D7\u30ED\u30D1\u30C6\u30A3\u3
 reason=\u7406\u7531
 reloadHost=Reload Host
 reloadServerGroup=\u30B5\u30FC\u30D0\u30FC\u30B0\u30EB\u30FC\u30D7\u306E\u518D\u30ED\u30FC\u30C9
-reloadServerNow=\u30B5\u30FC\u30D0\u30FC\u3092\u4ECA\u3059\u3050\u518D\u30ED\u30FC\u30C9\u3057\u3066\u304F\u3060\u3055\u3044\uFF01
+reloadServerNow=Reload server Now!
 remoteSocketDescription=\u30EA\u30E2\u30FC\u30C8\u5B9B\u5148\u306E\u30A2\u30A6\u30C8\u30D0\u30A6\u30F3\u30C9\u30BD\u30B1\u30C3\u30C8\u30D0\u30A4\u30F3\u30C7\u30A3\u30F3\u30B0\u306B\u95A2\u3059\u308B\u8A2D\u5B9A\u60C5\u5831\u3002
 removeProfile=\u30D7\u30ED\u30D5\u30A1\u30A4\u30EB\u306E\u524A\u9664
 replaceDeployment=\u30C7\u30D7\u30ED\u30A4\u30E1\u30F3\u30C8\u306E\u7F6E\u63DB
@@ -440,6 +441,7 @@ server_config_uptodate=\u30B5\u30FC\u30D0\u30FC\u8A2D\u5B9A\u306F\u6700\u65B0\u3
 server_instance_pleaseSelect=\u30B5\u30FC\u30D0\u30FC\u30A4\u30F3\u30B9\u30BF\u30F3\u30B9\u3092\u9078\u629E\u3057\u3066\u304F\u3060\u3055\u3044\u3002
 server_instance_reloadRequired=\u30B5\u30FC\u30D0\u30FC\u3092\u518D\u5EA6\u8AAD\u307F\u8FBC\u3093\u3067\u304F\u3060\u3055\u3044\u3002
 server_instance_servers_needReload=\u4EE5\u4E0B\u306E\u30B5\u30FC\u30D0\u30FC\u3092\u518D\u30ED\u30FC\u30C9\u3059\u308B\u5FC5\u8981\u304C\u3042\u308A\u307E\u3059\u3002
+server_instance_servers_needReload=The following servers need to be reloaded:
 server_instance_servers_needRestart=\u4EE5\u4E0B\u306E\u30B5\u30FC\u30D0\u30FC\u3092\u518D\u8D77\u52D5\u3059\u308B\u5FC5\u8981\u304C\u3042\u308A\u307E\u3059\u3002
 server_reload_desc=\u30B5\u30FC\u30D0\u30FC\u8A2D\u5B9A\u3092\u30EA\u30ED\u30FC\u30C9\u3059\u308B\u5FC5\u8981\u304C\u3042\u308A\u307E\u3059\u3002\u30EA\u30ED\u30FC\u30C9\u3057\u306A\u3044\u3068\u30B5\u30FC\u30D0\u30FC\u306F\u518D\u8D77\u52D5\u3055\u308C\u307E\u3059\u3002
 server_reload_title=\u30B5\u30FC\u30D0\u30FC\u8A2D\u5B9A\u306E\u518D\u8AAD\u307F\u8FBC\u307F
