+++ gui/src/main/java/org/jboss/as/console/client/core/UIConstants_pt_BR.properties
@@ -182,6 +182,7 @@ failedToReadMetricDescription=Falha ao recuperar as descri\u00E7\u00F5es m\u00E9
 failedToRemoveServer=Falha ao remover o servidor
 failedToResolveExpression=Falha ao resolver a express\u00E3o
 failedToRetrieveAttributeDescriptions=Falha ao recuperar as descri\u00E7\u00F5es de atributos.
+goToRuntime=Go to Runtime
 groupAlreadyExists=O grupo j\u00E1 existe
 help=Ajuda
 help_close_help=Encerrar
@@ -405,7 +406,7 @@ properties_global_desc=Essas propriedades est\u00E3o dispon\u00EDveis atrav\u00E
 reason=Raz\u00E3o
 reloadHost=Reload Host
 reloadServerGroup=Recarregar o Grupo dos Servidores
-reloadServerNow=Recarregue o(s) servidor(es) agora\!
+reloadServerNow=Reload server Now!
 remoteSocketDescription=A informa\u00E7\u00E3o da configura\u00E7\u00E3o para a destina\u00E7\u00E3o remota, de um socket binding de sa\u00EDda.
 removeProfile=Remover Perfil
 replaceDeployment=Substituir Implanta\u00E7\u00E3o
@@ -439,7 +440,8 @@ server_config_desc=O status da configura\u00E7\u00E3o do servidor. \u00C0s vezes
 server_config_uptodate=A configura\u00E7\u00E3o do servidor parece ter sido atualizada\!
 server_instance_pleaseSelect=Por favor, selecione a inst\u00E2ncia do servidor\:
 server_instance_reloadRequired=A recarga do servidor \u00E9 solicitada\!
-server_instance_servers_needReload=Os serguintes servidores precisam ser recarregados\:
+server_instance_servers_needReload=The following servers need to be reloaded:
+server_instance_servers_needReload_from_runtime=Os serguintes servidores precisam ser recarregados\:
 server_instance_servers_needRestart=Os seguintes servidores precisam ser reiniciados\:
 server_reload_desc=A configura\u00E7\u00E3o do servidor precisa ser recarregada ou o servidor reinicializado\!
 server_reload_title=Recarregamento da configura\u00E7\u00E3o do servidor
