+++ gui/src/main/java/org/jboss/as/console/client/core/UIConstants_zh_Hans.properties
@@ -182,6 +182,7 @@ failedToReadMetricDescription=\u83B7\u53D6\u5EA6\u91CF\u63CF\u8FF0\u5931\u8D25
 failedToRemoveServer=\u5220\u9664\u670D\u52A1\u5668\u5931\u8D25
 failedToResolveExpression=\u89E3\u6790\u8868\u8FBE\u5F0F\u5931\u8D25
 failedToRetrieveAttributeDescriptions=\u83B7\u53D6\u5C5E\u6027\u63CF\u8FF0\u5931\u8D25
+goToRuntime=Go to Runtime
 groupAlreadyExists=\u7EC4\u5DF2\u5B58\u5728
 help=\u5E2E\u52A9
 help_close_help=\u5173\u95ED
@@ -405,7 +406,7 @@ properties_global_desc=\u8FD9\u4E9B\u5C5E\u6027\u5728\u6574\u4E2A\u914D\u7F6E\u9
 reason=\u539F\u56E0
 reloadHost=Reload Host
 reloadServerGroup=\u91CD\u65B0\u52A0\u8F7D\u670D\u52A1\u5668\u7EC4
-reloadServerNow=\u73B0\u5728\u91CD\u65B0\u52A0\u8F7D\u670D\u52A1\u5668\uFF01
+reloadServerNow==Reload server Now!
 remoteSocketDescription=\u7528\u4E8E\u8FDC\u7A0B\u76EE\u7684\u5730\u3001\u8F6C\u51FA\u5957\u63A5\u5B57\u7ED1\u5B9A\u7684\u914D\u7F6E\u4FE1\u606F\u3002
 removeProfile=\u5220\u9664\u914D\u7F6E\u96C6
 replaceDeployment=\u66FF\u6362\u90E8\u7F72
@@ -439,7 +440,8 @@ server_config_desc=\u670D\u52A1\u5668\u914D\u7F6E\u72B6\u6001\u3002\u5728\u67D0\
 server_config_uptodate=\u670D\u52A1\u5668\u914D\u7F6E\u4F3C\u4E4E\u662F\u6700\u65B0\u7248\u672C\uFF01
 server_instance_pleaseSelect=\u8BF7\u9009\u62E9\u4E00\u4E2A\u670D\u52A1\u5668\u5B9E\u4F8B\uFF1A
 server_instance_reloadRequired=\u5FC5\u987B\u91CD\u65B0\u52A0\u8F7D\u670D\u52A1\u5668\uFF01
-server_instance_servers_needReload=\u4E0B\u5217\u670D\u52A1\u5668\u9700\u8981\u91CD\u65B0\u52A0\u8F7D\uFF1A
+server_instance_servers_needReload=The following servers need to be reloaded:
+server_instance_servers_needReload_from_runtime=\u4E0B\u5217\u670D\u52A1\u5668\u9700\u8981\u91CD\u65B0\u52A0\u8F7D\uFF1A
 server_instance_servers_needRestart=\u4E0B\u9762\u7684\u670D\u52A1\u5668\u9700\u8981\u91CD\u542F\uFF1A
 server_reload_desc=\u670D\u52A1\u5668\u914D\u7F6E\u9700\u8981\u91CD\u65B0\u52A0\u8F7D\u6216\u91CD\u542F\u670D\u52A1\u5668\uFF01
 server_reload_title=\u91CD\u65B0\u52A0\u8F7D\u670D\u52A1\u5668\u914D\u7F6E
