+++ gui/src/main/java/org/jboss/as/console/client/standalone/runtime/StandaloneRuntimePresenter.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import com.google.gwt.event.shared.GwtEvent;
 import com.google.gwt.safehtml.shared.SafeHtml;
-import com.google.gwt.user.client.rpc.AsyncCallback;
 import com.google.inject.Inject;
 import com.google.web.bindery.event.shared.EventBus;
 import com.gwtplatform.mvp.client.Presenter;
@@ -22,17 +21,14 @@ import org.jboss.as.console.client.Console;
 import org.jboss.as.console.client.core.Header;
 import org.jboss.as.console.client.core.MainLayoutPresenter;
 import org.jboss.as.console.client.core.NameTokens;
-import org.jboss.as.console.client.core.message.Message;
 import org.jboss.as.console.client.domain.model.SimpleCallback;
 import org.jboss.as.console.client.domain.model.SrvState;
 import org.jboss.as.console.client.domain.model.SuspendState;
 import org.jboss.as.console.client.semver.ManagementModel;
 import org.jboss.as.console.client.semver.Version;
-import org.jboss.as.console.client.shared.flow.TimeoutOperation;
 import org.jboss.as.console.client.shared.model.SubsystemLoader;
 import org.jboss.as.console.client.shared.model.SubsystemRecord;
-import org.jboss.as.console.client.shared.patching.RestartOp;
-import org.jboss.as.console.client.shared.schedule.LongRunningTask;
+import org.jboss.as.console.client.shared.runtime.StandaloneRestartReload;
 import org.jboss.as.console.client.shared.state.ReloadEvent;
 import org.jboss.as.console.client.shared.state.ReloadState;
 import org.jboss.as.console.client.shared.state.StandaloneRuntimeRefresh;
@@ -43,7 +39,6 @@ import org.jboss.as.console.client.widgets.nav.v3.PreviewEvent;
 import org.jboss.as.console.spi.RequiredResources;
 import org.jboss.ballroom.client.widgets.window.DefaultWindow;
 import org.jboss.dmr.client.ModelNode;
-import org.jboss.dmr.client.dispatch.AsyncCommand;
 import org.jboss.dmr.client.dispatch.DispatchAsync;
 import org.jboss.dmr.client.dispatch.impl.DMRAction;
 import org.jboss.dmr.client.dispatch.impl.DMRResponse;
@@ -55,13 +50,14 @@ import static org.jboss.dmr.client.ModelDescriptionConstants.*;
  */
 public class StandaloneRuntimePresenter
         extends Presenter<StandaloneRuntimePresenter.MyView, StandaloneRuntimePresenter.MyProxy>
-        implements Finder, PreviewEvent.Handler, FinderScrollEvent.Handler, StandaloneRuntimeRefresh.Handler {
+        implements Finder, PreviewEvent.Handler, FinderScrollEvent.Handler, StandaloneRuntimeRefresh.Handler, ReloadEvent.ReloadListener {
 
     private final PlaceManager placeManager;
     private final SubsystemLoader subsysStore;
     private final Header header;
     private final ReloadState reloadState;
     private final DispatchAsync dispatcher;
+    private final StandaloneRestartReload standaloneServerReload;
 
     private boolean hasBeenLoaded;
     private DefaultWindow window;
@@ -115,6 +111,7 @@ public class StandaloneRuntimePresenter
         this.header = header;
         this.reloadState = reloadState;
         this.dispatcher = dispatcher;
+        this.standaloneServerReload = new StandaloneRestartReload();
     }
 
     @Override
@@ -139,6 +136,7 @@ public class StandaloneRuntimePresenter
         getEventBus().addHandler(PreviewEvent.TYPE, this);
         getEventBus().addHandler(FinderScrollEvent.TYPE, this);
         getEventBus().addHandler(StandaloneRuntimeRefresh.TYPE, this);
+        getEventBus().addHandler(ReloadEvent.TYPE, this);
     }
 
     @Override
@@ -177,89 +175,7 @@ public class StandaloneRuntimePresenter
     }
 
     public void onReloadServerConfig() {
-        final ModelNode operation = new ModelNode();
-        operation.get(OP).set("reload");
-        operation.get(ADDRESS).setEmptyList();
-
-        dispatcher.execute(new DMRAction(operation), new AsyncCallback<DMRResponse>() {
-            @Override
-            public void onSuccess(DMRResponse result) {
-                ModelNode response = result.get();
-
-                if (response.isFailure()) {
-                    Console.error(Console.MESSAGES.failed("Reload Server"), response.getFailureDescription());
-                } else {
-                    pollState("Waiting for the server to reload","Reload Server");
-                }
-            }
-
-            @Override
-            public void onFailure(Throwable caught) {
-                Console.error(Console.MESSAGES.failed("Reload Server"), caught.getMessage());
-            }
-        });
-    }
-
-    private void pollState(final String standBy, final String success) {
-
-        LongRunningTask poll = new LongRunningTask(new AsyncCommand<Boolean>() {
-            @Override
-            public void execute(final AsyncCallback<Boolean> callback) {
-                checkServerState(callback,standBy,success);
-            }
-        }, 10);
-
-        // kick of the polling request
-        poll.schedule(500);
-    }
-
-    /**
-     * Simply query the process state attribute to get to the required headers
-     */
-    public void checkServerState(final AsyncCallback<Boolean> callback, final String standBy, final String success) {
-
-        // :read-attribute(name=process-type)
-        final ModelNode operation = new ModelNode();
-        operation.get(OP).set(READ_ATTRIBUTE_OPERATION);
-        operation.get(NAME).set("server-state");
-        operation.get(ADDRESS).setEmptyList();
-
-        dispatcher.execute(new DMRAction(operation), new SimpleCallback<DMRResponse>() {
-
-            @Override
-            public void onSuccess(DMRResponse result) {
-
-                ModelNode response = result.get();
-
-                if (response.isFailure()) {
-                    callback.onFailure(new RuntimeException("Failed to poll server state"));
-                } else {
-                    // TODO: only works when this response changes the reload state
-                    String outcome = response.get(RESULT).asString();
-                    boolean keepRunning = !outcome.equalsIgnoreCase("running");//reloadState.isStaleModel();
-
-                    if (!keepRunning) {
-
-                        // clear state
-                        reloadState.reset();
-
-                        Console.info(Console.MESSAGES.successful(success));
-
-                        // clear reload state
-                        getEventBus().fireEvent(new ReloadEvent());
-
-                        getView().updateServer(new StandaloneServer(SrvState.UNDEFINED, SuspendState.UNKOWN));
-                    }
-
-                    callback.onSuccess(keepRunning);
-                }
-            }
-
-            @Override
-            public void onFailure(Throwable caught) {
-                Console.getMessageCenter().notify(new Message(standBy, caught.getMessage(), Message.Severity.Warning));
-            }
-        });
+        new StandaloneRestartReload().onReloadServer();
     }
 
 
@@ -348,29 +264,14 @@ public class StandaloneRuntimePresenter
             }
         });
     }
-    
-	public void onRestartServer() {
-	    final ModelNode operation = new ModelNode();
-        operation.get(OP).set("shutdown");
-        operation.get(ADDRESS).setEmptyList();
-        operation.get("restart").set(true);
 
-        dispatcher.execute(new DMRAction(operation), new AsyncCallback<DMRResponse>() {
-            @Override
-            public void onSuccess(DMRResponse result) {
-                ModelNode response = result.get();
+    public void onRestartServer() {
+        new StandaloneRestartReload().onRestartServer();
+    }
 
-                if (response.isFailure()) {
-                    Console.error(Console.MESSAGES.failed("Restart Server"), response.getFailureDescription());
-                } else {
-                    pollState("Waiting for the server to restart","Restart Server");
-                }
-            }
+    @Override
+    public void onReload() {
+        getView().updateServer(new StandaloneServer(SrvState.UNDEFINED, SuspendState.UNKOWN));
+    }
 
-            @Override
-            public void onFailure(Throwable caught) {
-                Console.error(Console.MESSAGES.failed("Restart Server"), caught.getMessage());
-            }
-        });
-	}
 }
